refactor: debio conversion to common.
