+++ src/endpoints/ethereum-indexed-data/ethereum-indexed-data.module.ts
@@ -4,7 +4,7 @@ import { ServiceRequestController } from './service-request/service-request.cont
 import { ServiceRequestService } from './service-request/service-request.service';
 import { LocationModule } from '../../endpoints/location/location.module';
 import { EthereumModule } from '../../endpoints/ethereum/ethereum.module';
-import { DebioConversionModule } from '../../debio-conversion/debio-conversion.module';
+import { DebioConversionModule } from '../../common/debio-conversion/debio-conversion.module';
 
 @Module({
   imports: [
