+++ src/reward/reward.module.ts
@@ -1,6 +1,6 @@
 import { forwardRef, Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { DebioConversionModule } from '../debio-conversion/debio-conversion.module';
+import { DebioConversionModule } from '../common/debio-conversion/debio-conversion.module';
 
 import { SubstrateModule } from '../substrate/substrate.module';
 import { Reward } from './models/reward.entity';
